01__backend/requirements/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.6.2
cffi==1.15.1
click==8.1.5
cryptography==41.0.2
Flask==2.3.2
Flask-RESTful==0.3.10
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.3.6
